Added new AbstractEnrichBasedCrossClusterTestCase and all enrich based CCS tests are now using it
